change data types in delay functions

+++ lib/Delay.c
@@ -1,11 +1,11 @@
 #include "EmbVars.h"
 
-void Delay_us(u32 t) {
+void Delay_us(u64 t) {
 	for(u32 i = 0; i < t; i++){
 		asm("nop");
 	}
 }
 
-void Delay_ms(u16 t) {
+void Delay_ms(u32 t) {
 	Delay_us(t * 1000);
 }
